+++ internal/api/handler/transaction/add/contract.go
@@ -3,5 +3,5 @@ package add
 import "context"
 
 type usecaseTransactionAdd interface {
-	Add(ctx context.Context, categoryID int, amount float64) error
+	Add(ctx context.Context, categoryID int, amount int64) error
 }
