+++ internal/usecase/transaction/update/contract.go
@@ -3,5 +3,5 @@ package update
 import "context"
 
 type transactionRepo interface {
-	UpdateTransaction(ctx context.Context, id int, amount float64) error
+	UpdateTransaction(ctx context.Context, id int, amount int64) error
 }
